refactor: Managers폴더 제거 및 api, localStorage 폴더로 분리
